Construct mask for patch boundaries
We apply strong, homogeneous, boundary conditions to all dofs on the
boundary of patches.  We do not include facets on the boundary of
the *domain* as being on the boundary of patches.  If strong (Dirichlet)
conditions are supplied to the bilinear form, then these are used to
kill dofs on the domain boundary, otherwise we assume there are
Neumann-like conditions (and so the dofs should not be killed).